changed sockethandler to being websocketService
